third_party/WebKit/LayoutTests/fast/scroll-behavior/overflow-scroll-precise-deltas-dont-animate.html

<!DOCTYPE html>
<style>

#container {
  width: 200px;
  height: 200px;
  overflow: scroll;
}

#content {
  width: 750px;
  height: 10000px;
  background-color: blue;
}

</style>
<script src="../../resources/js-test.js"></script>
<div id="container">
  <div id="content"></div>
</div>
<div id="console"></div>
<script>

jsTestIsAsync = true;
element = document.getElementById("container");

var prevScrollTop;
var scrollStart;

function runTest() {
    internals.settings.setScrollAnimatorEnabled(true);

    element.addEventListener("scroll", onElementScroll);

    testRunner.layoutAndPaintAsyncThen(function() {
        // Give the container focus.
        eventSender.mouseMoveTo(100, 100);
        eventSender.mouseScrollBy(0, -20, /* paged */ false,
            /* preciseDeltas */ true);
    });
}

function onElementScroll() {
    shouldBe("element.scrollTop",
             "800");
    finishJSTest();
}

description("Tests that overflow scrolls with precise deltas aren't animated.");

if (window.eventSender)
    runTest();
else
    debug("FAIL: This test requires window.eventSender.");

</script>
