third_party/WebKit/LayoutTests/fast/scroll-behavior/overflow-scroll-precise-deltas-dont-animate-expected.txt

Tests that overflow scrolls with precise deltas aren't animated.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.scrollTop is 800
PASS successfullyParsed is true

TEST COMPLETE
